Newer API and removal of Burn.
